# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AiAgentEnabledRequest": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "enabled"
        ],
        "title": "AiAgentEnabledRequest",
        "type": "object"
      },
...
      "AiAgentSettingsResponse": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AiAgentResponse"
            },
            "title": "Agents",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "agents"
        ],
        "title": "AiAgentSettingsResponse",
        "type": "object"
      },
      "AiAgentUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "api_base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "maxLength": 65536,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_protocol": {
            "$ref": "#/components/schemas/ModelApiProtocol",
            "default": "chat_completions"
          },
          "clear_api_key": {
            "default": false,
            "title": "Clear Api Key",
            "type": "boolean"
          },
          "connect_timeout_seconds": {
            "default": 5.0,
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Connect Timeout Seconds",
            "type": "number"
          },
          "context_window_tokens": {
            "default": 128000,
            "maximum": 4000000.0,
            "minimum": 8192.0,
            "title": "Context Window Tokens",
            "type": "integer"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "max_batch_input_tokens": {
            "default": 64000,
            "maximum": 4000000.0,
            "minimum": 4096.0,
            "title": "Max Batch Input Tokens",
            "type": "integer"
          },
          "max_output_tokens": {
            "default": 8192,
            "maximum": 131072.0,
            "minimum": 256.0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "max_retries": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Max Retries",
            "type": "integer"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "pool_timeout_seconds": {
            "default": 5.0,
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Pool Timeout Seconds",
            "type": "number"
          },
          "provider": {
            "$ref": "#/components/schemas/ModelProvider"
          },
          "read_timeout_seconds": {
            "default": 180.0,
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Read Timeout Seconds",
            "type": "number"
          },
          "reasoning_effort": {
            "$ref": "#/components/schemas/ModelReasoningEffort",
            "default": "none"
          },
          "write_timeout_seconds": {
            "default": 30.0,
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Write Timeout Seconds",
            "type": "number"
          }
        },
        "required": [
          "expected_revision",
          "provider",
          "model"
        ],
        "title": "AiAgentUpdateRequest",
        "type": "object"
      },
...
      "AiProviderUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "api_base_url": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "maxLength": 65536,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "api_protocol": {
            "$ref": "#/components/schemas/ModelApiProtocol"
          },
          "cache_read_usd_per_million": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Read Usd Per Million"
          },
          "cache_write_usd_per_million": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Write Usd Per Million"
          },
          "clear_api_key": {
            "default": false,
            "title": "Clear Api Key",
            "type": "boolean"
          },
          "connect_timeout_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Connect Timeout Seconds",
            "type": "number"
          },
          "context_window_tokens": {
            "maximum": 4000000.0,
            "minimum": 8192.0,
            "title": "Context Window Tokens",
            "type": "integer"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "input_usd_per_million": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Usd Per Million"
          },
          "max_batch_input_tokens": {
            "default": 64000,
            "maximum": 4000000.0,
            "minimum": 4096.0,
            "title": "Max Batch Input Tokens",
            "type": "integer"
          },
          "max_output_tokens": {
            "maximum": 131072.0,
            "minimum": 256.0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "max_request_bytes": {
            "maximum": 10485760.0,
            "minimum": 65536.0,
            "title": "Max Request Bytes",
            "type": "integer"
          },
          "max_response_bytes": {
            "maximum": 10485760.0,
            "minimum": 65536.0,
            "title": "Max Response Bytes",
            "type": "integer"
          },
          "model": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Model",
            "type": "string"
          },
          "output_usd_per_million": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,6}0*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Usd Per Million"
          },
          "pool_timeout_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Pool Timeout Seconds",
            "type": "number"
          },
          "read_timeout_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Read Timeout Seconds",
            "type": "number"
          },
          "reasoning_effort": {
            "$ref": "#/components/schemas/ModelReasoningEffort",
            "default": "none"
          },
          "write_timeout_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 3600.0,
            "title": "Write Timeout Seconds",
            "type": "number"
          }
        },
        "required": [
          "expected_revision",
          "model",
          "api_protocol",
          "context_window_tokens",
          "max_output_tokens",
          "connect_timeout_seconds",
          "read_timeout_seconds",
          "write_timeout_seconds",
          "pool_timeout_seconds",
          "max_request_bytes",
          "max_response_bytes"
        ],
        "title": "AiProviderUpdateRequest",
        "type": "object"
      },
      "AiRevisionRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "AiRevisionRequest",
        "type": "object"
      },
...
      "ConfigurationAuditListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConfigurationAuditResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ConfigurationAuditListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelApiProtocol": {
        "description": "\u4f9b\u5e94\u5546\u8c03\u7528\u4f7f\u7528\u7684\u7a33\u5b9a HTTP API \u534f\u8bae\u3002",
        "enum": [
          "responses",
          "chat_completions",
          "messages"
        ],
        "title": "ModelApiProtocol",
        "type": "string"
      },
      "ModelProvider": {
        "description": "\u6a21\u578b\u8c03\u7528\u9002\u914d\u5668\u652f\u6301\u7684\u4f9b\u5e94\u5546\u3002",
        "enum": [
          "openai",
          "anthropic"
        ],
        "title": "ModelProvider",
        "type": "string"
      },
...
      "ReviewAgent": {
        "description": "\u56fa\u5b9a DAG \u4e2d\u7684\u56db\u4e2a\u6a21\u578b\u8282\u70b9\u3002",
        "enum": [
          "security",
          "convention",
          "logic",
          "summary"
        ],
        "title": "ReviewAgent",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "OpenReviewer API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings/ai/agents/{agent}": {
      "put": {
        "operationId": "update_agent_settings_api_v1_settings_ai_agents__agent__put",
        "parameters": [
          {
            "in": "path",
            "name": "agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReviewAgent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiAgentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiAgentSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent Settings"
      }
    },
    "/api/v1/settings/ai/agents/{agent}/enabled": {
      "post": {
        "operationId": "set_agent_enabled_api_v1_settings_ai_agents__agent__enabled_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReviewAgent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiAgentEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiAgentSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Agent Enabled"
      }
    },
    "/api/v1/settings/ai/agents/{agent}/test": {
      "post": {
        "operationId": "test_agent_settings_api_v1_settings_ai_agents__agent__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ReviewAgent"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiAgentSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Agent Settings"
      }
    },
    "/api/v1/settings/ai/providers/{provider}": {
      "put": {
        "operationId": "update_ai_provider_api_v1_settings_ai_providers__provider__put",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModelProvider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiProviderUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Ai Provider"
      }
    },
    "/api/v1/settings/ai/providers/{provider}/activate": {
      "post": {
        "operationId": "activate_ai_provider_api_v1_settings_ai_providers__provider__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModelProvider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Ai Provider"
      }
    },
    "/api/v1/settings/ai/providers/{provider}/test": {
      "post": {
        "operationId": "test_ai_provider_api_v1_settings_ai_providers__provider__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ModelProvider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Ai Provider"
      }
    },
    "/api/v1/settings/ai/review-policy": {
      "put": {
        "operationId": "update_review_policy_api_v1_settings_ai_review_policy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPolicyUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Review Policy"
      }
    },
    "/api/v1/settings/audits": {
      "get": {
        "operationId": "list_configuration_audits_api_v1_settings_audits_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationAuditListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Configuration Audits"
      }
    },
...
  }
}